feat:  알림장 스켈레톤 수정 #77
