+++ src/components/notification/NotificationCard.tsx
@@ -1,6 +1,4 @@
 import { useCallback } from 'react';
-import { Badge } from '@/components/ui/badge';
-import { Button } from '@/components/ui/button';
 
 interface NotificationCardProps {
   title: string;
